Giving ProjectList some loving <3
