+++ SharePointFramework/PortfolioWebParts/src/components/ProjectList/ProjectList.module.scss
@@ -26,6 +26,11 @@
             @include ms-sm12;
             margin: 15px 0 0 0;
         }
+        .spinner {
+            @include ms-Grid-col;
+            @include ms-sm12;
+            margin: 15px 0 0 0;
+        }
         .projects {
             @include ms-Grid-col;
             @include ms-sm12;
