+++ SharePointFramework/PortfolioWebParts/src/loc/nb-no.js
@@ -104,7 +104,7 @@ define([], function() {
         NoGroupingText: 'Ingen gruppering',
         NoProjectData: 'Kan ikke hente alle data fra prosjektområdet. Det kan være at du ikke har tilgang til selve området. Det kan også være at området nylig er opprettet eller at prosjektegenskaper ikke er utfylt.',
         NoProperties: 'Det er ikke satt noen egenskaper for prosjektet',
-        NoSearchResults: 'Søket ga ingen resultater',
+        ProjectListEmptyText: 'Fant ingen prosjekter å vise.',
         NotSet: 'Ikke satt',
         PhaseLabel: 'Fase',
         ProbabilityLabel: 'Sannsynlighet (S)',
